+++ package.json
@@ -53,6 +53,7 @@
   },
   "devDependencies": {
     "@edx/frontend-build": "^5.5.1",
+    "@testing-library/react": "10.3.0",
     "axios-mock-adapter": "^1.19.0",
     "check-prop-types": "1.1.2",
     "codecov": "3.8.1",
@@ -63,7 +64,6 @@
     "jest": "26.6.3",
     "react-test-renderer": "17.0.1",
     "reactifex": "1.1.1",
-    "request": "^2.88.2",
-    "@testing-library/react": "10.3.0"
+    "request": "^2.88.2"
   }
 }
