+++ src/index.scss
@@ -4,7 +4,7 @@
 @import "~@edx/brand/paragon/overrides";
 @import '~@edx/paragon/scss/edx/theme.scss';
 
-@import "./support-header/index.scss";
+@import "./supportHeader/index.scss";
 
 @import "./users/index.scss";
