+++ src/supportHeader/Menu/menu.scss
@@ -1,6 +1,7 @@
 .menu {
   position: relative;
 }
+
 .menu-content {
   position: absolute;
   top: 100%;
@@ -15,19 +16,17 @@
   }
 }
 
-
 .menu-dropdown-enter {
   opacity: 0;
   transform-origin: 75% 0;
   transform: scale3d(0.8, 0.8, 1);
 }
+
 .menu-dropdown-enter-active {
   transform-origin: 75% 0;
   // transition: all 0ms cubic-bezier(0.4, 0, 0.2, 1);
   transform: scale3d(1, 1, 1);
   opacity: 1;
-}
-.menu-dropdown-enter-done {
 }
 
 .menu-dropdown-exit {
@@ -35,11 +34,10 @@
   transform: scale3d(1, 1, 1);
   opacity: 1;
 }
+
 .menu-dropdown-exit-active {
   transform-origin: 75% 0;
   transform: scale3d(0.8, 0.8, 1);
   // transition: all 0ms cubic-bezier(0.8, 0, 0.6, 1);
   opacity: 0;
-}
-.menu-dropdown-exit-done {
 }
